ACTIN-68: Move files in ckbtrial source to ckb source
