+++ algo/src/main/java/com/hartwig/serve/sources/ckb/ActionableTrial.java
@@ -1,4 +1,4 @@
-package com.hartwig.serve.sources.ckbtrial;
+package com.hartwig.serve.sources.ckb;
 
 import com.hartwig.serve.datamodel.ActionableEvent;
